Extract inspector-factory-compiler annotations to separate artifact
For android gradle plugin compatibility
