+++ inspector-factory-compiler-annotations/README.md
@@ -1,0 +1,12 @@
+inspector-factory-compiler-annotations
+======================================
+
+This artifact contains necessary compile-time annotations for the `inspector-factory-compiler`. 
+These are extracted so you can use them as a `compileOnly` dependency in Android projects without 
+placing the full annotation processor in the factory compiler artifact on your classpath.
+
+See the `inspector-factory-compiler` README for full documentation.
+
+```gradle
+implementation 'io.sweers.inspector:inspector-factory-compiler-annotations:x.y.z'
+```
