+++ inspector-factory-compiler-annotations/build.gradle
@@ -1,0 +1,21 @@
+buildscript {
+  repositories {
+    jcenter()
+    maven { url deps.build.repositories.plugins }
+  }
+  dependencies {
+    classpath deps.build.gradlePlugins.errorProne
+  }
+}
+
+apply plugin: 'java-library'
+apply plugin: 'net.ltgt.errorprone'
+
+sourceCompatibility = "1.8"
+targetCompatibility = "1.8"
+
+dependencies {
+  api project(':inspector')
+}
+
+apply from: rootProject.file('gradle/gradle-mvn-push.gradle')
