+++ inspector-factory-compiler-annotations/gradle.properties
@@ -1,0 +1,3 @@
+POM_NAME=Inspector Factory Compiler
+POM_ARTIFACT_ID=inspector-factory-compiler
+POM_PACKAGING=jar
