+++ inspector-sample/src/main/java/io/sweers/inspector/sample/SampleFactory.java
@@ -2,7 +2,7 @@ package io.sweers.inspector.sample;
 
 import com.google.auto.value.AutoValue;
 import io.sweers.inspector.Validator;
-import io.sweers.inspector.factorycompiler.InspectorFactory;
+import io.sweers.inspector.InspectorFactory;
 
 @InspectorFactory(include = AutoValue.class) public abstract class SampleFactory
     implements Validator.Factory {
